商品購入機能#10

+++ app/controllers/items_controller.rb
@@ -24,15 +24,15 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if current_user.id == item.user_id
+    if current_user.id == item.user_id && item.order.blank?
       item.destroy
       redirect_to root_path
     end
   end
 
   def edit
-    # ログイン状態の場合でも、自身が出品した売却済み商品の商品情報編集ページへ遷移しようとすると、トップページに遷移すること。は商品購入機能実装後に行う。
-    redirect_to root_path unless current_user.id == @item.user_id
+    # ①自身が出品した売却済み商品の商品情報編集ページへ遷移しようとする or ②出品者と異なるユーザーが遷移しようとすると、トップページに遷移する
+    redirect_to root_path if @item.order.present? || !(current_user.id == @item.user_id)
   end
 
   def update
